pyproject.toml
[project]
name = "story-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib==1.5.2",
    "fastapi==0.115.12",
    "google-auth==2.39.0",
    "google-cloud-datastore==2.21.0",
    "google-cloud-run==0.10.17",
    "google-cloud-storage==3.1.0",
    "google-genai==1.13.0",
    "gradio==5.47.2",
    "gradio-client==1.13.3",
    "itsdangerous==2.2.0",
    "mediapy>=1.2.4",
    "moviepy==2.1.2",
    "pydantic==2.11.4",
    "python-dotenv==1.1.0",
    "starlette==0.46.2",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
]
